Finalize

+++ src/components/auth/edit/DeleteUserAccount.js
@@ -1,17 +1,29 @@
 import { deleteUser } from "firebase/auth";
 import React from "react";
 import { auth } from "../../../utils/firebase";
-import { Navigate } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 const DeleteUserAccount = () => {
+	const navigate = useNavigate();
 	const deleteUserAcc = async () => {
-		deleteUser(auth.currentUser);
-		return <Navigate to='/' />;
+		const res = prompt(
+			"Are you sure you want to delete this account? (y/n)"
+		);
+
+		if (res.toLowerCase() === "n") {
+			deleteUser(auth.currentUser);
+			navigate("/");
+		}
 	};
 
 	return (
 		<div>
-			<h5>Delete Account</h5>
-			<button onClick={deleteUserAcc}>Delete Account</button>
+			<h5 className='font-bold text-lg mb-3'>Delete Account</h5>
+			<button
+				className='w-full inline-flex justify-center rounded-md border border-transparent shadow-sm px-4 py-2 bg-red-600 text-base font-medium text-white hover:bg-red-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-red-500  sm:w-auto sm:text-sm'
+				onClick={deleteUserAcc}
+			>
+				Delete Account
+			</button>
 		</div>
 	);
 };
